Backend/requirements.txt
aiohttp==3.7.4
aiosnmp==0.4.0
alabaster==0.7.12
async-timeout==3.0.1
asyncio==3.4.3
attrs==21.2.0
autopep8==1.5.7
Babel==2.9.1
certifi==2020.12.5
chardet==4.0.0
click==8.0.0
docutils==0.16
Flask==2.0.0
Flask-Cors==3.0.10
gevent==21.1.2
greenlet==1.1.0
grequests==0.6.0
hurry.filesize==0.9
idna==2.10
imagesize==1.2.0
itsdangerous==2.0.1
Jinja2==2.11.3
MarkupSafe==1.1.1
multidict==5.1.0
mysql-connector==2.2.9
packaging==20.9
pycodestyle==2.7.0
Pygments==2.9.0
pyparsing==2.4.7
python3-nmap==1.5.0
pytz==2021.1
pyzabbix==0.8.2
requests==2.25.1
simplejson==3.17.2
six==1.16.0
snowballstemmer==2.1.0
Sphinx==4.0.1
sphinx-rtd-theme==0.5.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.4
Werkzeug==2.0.1
yarl==1.6.3
zope.event==4.5.0
zope.interface==5.4.0
